requirements.txt
streamlit
python-dotenv
langchain_tavily
langchain_groq
langchain_core
requests
beautifulsoup4

